CustomerBase State Refactoring
Goal was to control the customer based on state as much as possible due to consistency issues in network play.

The idea is you set the customer state and then other stuff happens.

Example: Instead of invoking Leaving() with a delay on Order(), when the customer is in "InSit" state, it checks if their timer threshold has been met and eventually changes to "Leaving" state

- Changed Update() to only contain a switch for state
- Created Update<action>() for each state change
- Removed dead "Usings"
- Fixed bug where customer review doesn't appear
- Removed the need for GetCustomerReview() from orderstats